Auto-refresh the browser on version update
Also displays latest version in the header.

The application reads the current version from the package.json file and returns this to any browser that requests it via a web socket.

The browser on load, emits an event requesting the current version.  It then continuously emits events to check for the latest version.  If it changes then the browser is refreshed.